Don't incorporate end date into open-ended query caches
Introduces a class to manage an index of the query caches for these
queries, so that they can also be invalidated as needed later. The
query class now adds the query cache to the index if the query is open-
ended.

The network queries needs to be indexed at the network level, so the
method to determine if a query is a network query up to the main query
class.

See #7