Flusher: support only flushing matching caches
If the user ID isn't among the cached results, then we don't flush the
cache in this case.

This optimization will be useful when we are flushing the caches for a
user after they are deleted, since we know that if they aren't in the
list of results the cache is still good. In other cases, we don't know
that, because the user might be moving into a top spot even if they
don't hold one already.

See #12