Query: use 'offset' arg instead of 'start'
The 'start' query arg has been deprecated in WordPoints because it is
actually an offset. This commit updates our code to avoid using that
arg, using the new 'offset' arg instead.

The 'start' arg was not used in block signatures, so the blocks are
still good. The arg was used in the query results cache signatures,
however, so most caches will have to be regenerated.

See WordPoints/wordpoints#688